awwvision/worker/src/requirements.txt
gcloud==0.8.0
google-api-python-client==1.6.3
google-cloud-vision==0.30.0
oauth2client==4.1.2
redis==2.10.5
requests[security]==2.8.1
psq==0.2.1
